Lighting maps 02

+++ opengl-project/shader.cpp
@@ -74,11 +74,11 @@ void Shader::setBool(const std::string &name, bool value) const
 };
 void Shader::setInt(const std::string &name, int value) const
 {
-  glUniform1i(glGetUniformLocation(ID, name.c_str()), (int)value);
+  glUniform1i(glGetUniformLocation(ID, name.c_str()), value);
 };
 void Shader::setFloat(const std::string &name, float value) const
 {
-  glUniform1i(glGetUniformLocation(ID, name.c_str()), (float)value);
+  glUniform1i(glGetUniformLocation(ID, name.c_str()), value);
 };
 void Shader::setMat4(const std::string &name, const glm::mat4 &mat) const
 {
